requirements.txt
Django==3.0.8
psycopg2-binary
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-rest-auth
django-allauth
django-extensions
django_filter
django-enumchoicefield
django-extra-fields
Pillow==8.0.1
wdb